move variantgetters to its own helper file
